Move SqlParameterization
