+++ src/Microsoft.SqlTools.LanguageService/AutoParameterizaition/Exceptions/ParameterizationParsingException.cs
@@ -7,7 +7,7 @@
 
 using System;
 
-namespace Microsoft.SqlTools.ServiceLayer.AutoParameterizaition.Exceptions
+namespace Microsoft.SqlTools.LanguageService.AutoParameterizaition.Exceptions
 {
     /// <summary>
     /// ParameterizationParsingException is used to surface parse errors encountered in the TSQL batch while creating a parse tree
