+++ src/Microsoft.SqlTools.LanguageService/AutoParameterizaition/SqlParameterizer.cs
@@ -10,10 +10,10 @@ using System.Data.Common;
 using System.IO;
 using System.Linq;
 using Microsoft.SqlServer.TransactSql.ScriptDom;
-using Microsoft.SqlTools.ServiceLayer.AutoParameterizaition.Exceptions;
+using Microsoft.SqlTools.LanguageService.AutoParameterizaition.Exceptions;
 using Microsoft.SqlTools.LanguageService.Workspace.Contracts;
 
-namespace Microsoft.SqlTools.ServiceLayer.AutoParameterizaition
+namespace Microsoft.SqlTools.LanguageService.AutoParameterizaition
 {
     public static class SqlParameterizer
     {
