+++ src/Microsoft.SqlTools.LanguageService/Localization/sr.cs
@@ -163,6 +163,41 @@ namespace Microsoft.SqlTools.LanguageService
             return Keys.GetString(Keys.PeekDefinitionError, errorMessage);
         }
 
+        public static string ParameterizationDetails(string variableName, string sqlDbType, int size, int precision, int scale, string sqlValue)
+        {
+            return Keys.GetString(Keys.ParameterizationDetails, variableName, sqlDbType, size, precision, scale, sqlValue);
+        }
+
+        public static string ErrorMessageHeader(int lineNumber)
+        {
+            return Keys.GetString(Keys.ErrorMessageHeader, lineNumber);
+        }
+
+        public static string ErrorMessage(string variableName, string sqlDataType, string literalValue)
+        {
+            return Keys.GetString(Keys.ErrorMessage, variableName, sqlDataType, literalValue);
+        }
+
+        public static string DateTimeErrorMessage(string variableName, string sqlDataType, string literalValue)
+        {
+            return Keys.GetString(Keys.DateTimeErrorMessage, variableName, sqlDataType, literalValue);
+        }
+
+        public static string BinaryLiteralPrefixMissingError(string variableName, string sqlDataType, string literalValue)
+        {
+            return Keys.GetString(Keys.BinaryLiteralPrefixMissingError, variableName, sqlDataType, literalValue);
+        }
+
+        public static string ParsingErrorHeader(int lineNumber, int columnNumber)
+        {
+            return Keys.GetString(Keys.ParsingErrorHeader, lineNumber, columnNumber);
+        }
+
+        public static string ScriptTooLarge(int maxChars, int currentChars)
+        {
+            return Keys.GetString(Keys.ScriptTooLarge, maxChars, currentChars);
+        }
+
         [System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
         public class Keys
         {
@@ -234,6 +269,27 @@ namespace Microsoft.SqlTools.LanguageService
             public const string PeekDefinitionTypeNotSupportedError = "PeekDefinitionTypeNotSupportedError";
 
 
+            public const string ParameterizationDetails = "ParameterizationDetails";
+
+
+            public const string ErrorMessageHeader = "ErrorMessageHeader";
+
+
+            public const string ErrorMessage = "ErrorMessage";
+
+
+            public const string DateTimeErrorMessage = "DateTimeErrorMessage";
+
+
+            public const string BinaryLiteralPrefixMissingError = "BinaryLiteralPrefixMissingError";
+
+
+            public const string ParsingErrorHeader = "ParsingErrorHeader";
+
+
+            public const string ScriptTooLarge = "ScriptTooLarge";
+
+
             private Keys()
             { }
 
@@ -261,6 +317,12 @@ namespace Microsoft.SqlTools.LanguageService
             }
 
 
+            public static string GetString(string key, object arg0, object arg1)
+            {
+                return string.Format(global::System.Globalization.CultureInfo.CurrentCulture, resourceManager.GetString(key, _culture), arg0, arg1);
+            }
+
+
             public static string GetString(string key, object arg0, object arg1, object arg2)
             {
                 return string.Format(global::System.Globalization.CultureInfo.CurrentCulture, resourceManager.GetString(key, _culture), arg0, arg1, arg2);
@@ -270,6 +332,12 @@ namespace Microsoft.SqlTools.LanguageService
             public static string GetString(string key, object arg0, object arg1, object arg2, object arg3)
             {
                 return string.Format(global::System.Globalization.CultureInfo.CurrentCulture, resourceManager.GetString(key, _culture), arg0, arg1, arg2, arg3);
+            }
+
+
+            public static string GetString(string key, object arg0, object arg1, object arg2, object arg3, object arg4, object arg5)
+            {
+                return string.Format(global::System.Globalization.CultureInfo.CurrentCulture, resourceManager.GetString(key, _culture), arg0, arg1, arg2, arg3, arg4, arg5);
             }
 
         }
