+++ src/Microsoft.SqlTools.ServiceLayer/Localization/sr.cs
@@ -10767,41 +10767,6 @@ namespace Microsoft.SqlTools.ServiceLayer
             return Keys.GetString(Keys.QueryServiceSaveAsFail, fileName, message);
         }
 
-        public static string ParameterizationDetails(string variableName, string sqlDbType, int size, int precision, int scale, string sqlValue)
-        {
-            return Keys.GetString(Keys.ParameterizationDetails, variableName, sqlDbType, size, precision, scale, sqlValue);
-        }
-
-        public static string ErrorMessageHeader(int lineNumber)
-        {
-            return Keys.GetString(Keys.ErrorMessageHeader, lineNumber);
-        }
-
-        public static string ErrorMessage(string variableName, string sqlDataType, string literalValue)
-        {
-            return Keys.GetString(Keys.ErrorMessage, variableName, sqlDataType, literalValue);
-        }
-
-        public static string DateTimeErrorMessage(string variableName, string sqlDataType, string literalValue)
-        {
-            return Keys.GetString(Keys.DateTimeErrorMessage, variableName, sqlDataType, literalValue);
-        }
-
-        public static string BinaryLiteralPrefixMissingError(string variableName, string sqlDataType, string literalValue)
-        {
-            return Keys.GetString(Keys.BinaryLiteralPrefixMissingError, variableName, sqlDataType, literalValue);
-        }
-
-        public static string ParsingErrorHeader(int lineNumber, int columnNumber)
-        {
-            return Keys.GetString(Keys.ParsingErrorHeader, lineNumber, columnNumber);
-        }
-
-        public static string ScriptTooLarge(int maxChars, int currentChars)
-        {
-            return Keys.GetString(Keys.ScriptTooLarge, maxChars, currentChars);
-        }
-
         public static string WritingServerContextualizationToCacheError(string message)
         {
             return Keys.GetString(Keys.WritingServerContextualizationToCacheError, message);
@@ -11194,27 +11159,6 @@ namespace Microsoft.SqlTools.ServiceLayer
 
 
             public const string SqlCmdUnsupportedToken = "SqlCmdUnsupportedToken";
-
-
-            public const string ParameterizationDetails = "ParameterizationDetails";
-
-
-            public const string ErrorMessageHeader = "ErrorMessageHeader";
-
-
-            public const string ErrorMessage = "ErrorMessage";
-
-
-            public const string DateTimeErrorMessage = "DateTimeErrorMessage";
-
-
-            public const string BinaryLiteralPrefixMissingError = "BinaryLiteralPrefixMissingError";
-
-
-            public const string ParsingErrorHeader = "ParsingErrorHeader";
-
-
-            public const string ScriptTooLarge = "ScriptTooLarge";
 
 
             public const string WritingServerContextualizationToCacheError = "WritingServerContextualizationToCacheError";
