+++ src/Microsoft.SqlTools.ServiceLayer/Localization/sr.resx
@@ -353,41 +353,6 @@
     <value>Encountered unsupported token {0}</value>
     <comment></comment>
   </data>
-  <data name="ParameterizationDetails" xml:space="preserve">
-    <value>{0} will be converted to a Microsoft.Data.SqlClient.SqlParameter object with the following properties: SqlDbType = {1}, Size = {2}, Precision = {3}, Scale = {4}, SqlValue = {5}</value>
-    <comment>.
- Parameters: 0 - variableName (string), 1 - sqlDbType (string), 2 - size (int), 3 - precision (int), 4 - scale (int), 5 - sqlValue (string) </comment>
-  </data>
-  <data name="ErrorMessageHeader" xml:space="preserve">
-    <value>Line {0}</value>
-    <comment>.
- Parameters: 0 - lineNumber (int) </comment>
-  </data>
-  <data name="ErrorMessage" xml:space="preserve">
-    <value>Unable to convert {0} to a Microsoft.Data.SqlClient.SqlParameter object. The specified literal cannot be converted to {1}(Microsoft.Data.SqlDbType). Literal value: {2}</value>
-    <comment>.
- Parameters: 0 - variableName (string), 1 - sqlDataType (string), 2 - literalValue (string) </comment>
-  </data>
-  <data name="DateTimeErrorMessage" xml:space="preserve">
-    <value>Unable to convert {0} to a Microsoft.Data.SqlClient.SqlParameter object. The specified literal cannot be converted to {1}(Microsoft.Data.SqlDbType), as it used an unsupported date/time format. Use one of the supported date/time formats. Literal value: {2}</value>
-    <comment>.
- Parameters: 0 - variableName (string), 1 - sqlDataType (string), 2 - literalValue (string) </comment>
-  </data>
-  <data name="BinaryLiteralPrefixMissingError" xml:space="preserve">
-    <value>Unable to convert {0} to a Microsoft.Data.SqlClient.SqlParameter object. The specified literal cannot be converted to {1}(Microsoft.Data.SqlDbType), as prefix 0x is expected for a binary literals.  Literal value: {2}</value>
-    <comment>.
- Parameters: 0 - variableName (string), 1 - sqlDataType (string), 2 - literalValue (string) </comment>
-  </data>
-  <data name="ParsingErrorHeader" xml:space="preserve">
-    <value>Line {0}, column {1}</value>
-    <comment>.
- Parameters: 0 - lineNumber (int), 1 - columnNumber (int) </comment>
-  </data>
-  <data name="ScriptTooLarge" xml:space="preserve">
-    <value>The current script is too large for Parameterization for Always Encrypted, please disable Parameterization for Always Encrypted in Query Options (Query &gt; Query Options &gt; Execution &gt; Advanced). Maximum allowable length: {0} characters, Current script length: {1} characters</value>
-    <comment>.
- Parameters: 0 - maxChars (int), 1 - currentChars (int) </comment>
-  </data>
   <data name="WritingServerContextualizationToCacheError" xml:space="preserve">
     <value>An error was encountered while writing server contextualization scripts to the cache. Error: {0}</value>
     <comment>.
