+++ src/Microsoft.SqlTools.ServiceLayer/QueryExecution/Batch.cs
@@ -21,7 +21,7 @@ using Microsoft.SqlTools.Utility;
 using System.Globalization;
 using Microsoft.SqlTools.ServiceLayer.Connection;
 using Microsoft.SqlTools.ServiceLayer.BatchParser;
-using Microsoft.SqlTools.ServiceLayer.AutoParameterizaition;
+using Microsoft.SqlTools.LanguageService.AutoParameterizaition;
 using Microsoft.SqlTools.LanguageService.Workspace;
 using Microsoft.SqlTools.ServiceLayer.SqlContext;
